+++ framework/wazuh/core/cluster/control.py
@@ -4,11 +4,10 @@
 import json
 
 from wazuh import common
-from wazuh.core.core_agent import Agent
 from wazuh.core.cluster import local_client
 from wazuh.core.cluster.common import as_wazuh_object, WazuhJSONEncoder
+from wazuh.core.core_agent import Agent
 from wazuh.utils import filter_array_by_query
-from wazuh.rbac.decorators import expose_resources
 
 
 async def get_nodes(lc: local_client.LocalClient, filter_node=None, offset=0, limit=common.database_limit,
